Revert "DRAFT: Suitable Web APIs supported"
