+++ .npmignore
@@ -1,7 +1,10 @@
-.github/
 node_modules/
-benchmark/
-test/
 tsdist/
+.github/
+benchmark.css
+index.html
+benchmark.js
+serve.js
+test.js
 README.md
 LICENSE
