+++ index.html
@@ -6,9 +6,8 @@
     <!-- https://stackoverflow.com/a/13416784/22334683 -->
     <link rel="icon" href="data:;base64,=">
     
-    <link rel="stylesheet" href="./benchmark/benchmark.css">
-    <script defer type="module" src="./benchmark/benchmark.js"></script>
-    
+    <link rel="stylesheet" href="./benchmark.css">
+    <script defer type="module" src="./benchmark.js"></script>
 </head>
 <body>
     <div class="main-container">
