+++ index.js
@@ -1,9 +1,4 @@
-import cloneDeep from "./src/clone-deep.js";
-import cloneDeepFully from "./src/clone-deep-fully.js";
-import useCustomizers from "./src/use-customizers.js";
-
+import cloneDeep from "./clone-deep.js";
+export { cloneDeepFully, useCustomizers } from "./clone-deep-utils.js";
+export { CLONE } from "./clone-deep-helpers.js";
 export default cloneDeep;
-export const cloneDeep = cloneDeep;
-export const cloneDeepFully = cloneDeepFully;
-export const useCustomizers = useCustomizers;
-export { CLONE } from "./src/utils.js";
